Added Email Library Functionality
With this update the email library should be able to be used to send emails and to store email entries in the database. The database could either be cosmodb or sqlserver and that is determined by the appsettings of the module that uses the library. Logging has also been added for all crucial funtionality. Testing and integration with the REST Email REST API will be done in a future commit.