Added Endpoints In Email Library REST API
I added all the necessary for now endpoints in the EmailLibraryAPI, which will enable the functionality of the EmailLibrary(CRUD on email entries and sending emails). In development there is swagger api documentation and appropriate error codes are sent for different possible and edge cases. Integration tests will be added in the next commit.