Added Authentication Operations In The Authentication Library
I added all the necessary operations for the typical authentication procedures such as sign in, sign up, sign out, reset password. I also added account change functionality in the form of update email, update password, delete account and others. This implementation is intended to work with JWTs and the API that will be implemented for this library should make sure that it is combatible. Logging has been also added for all available methods. The AppUser class for now that inheritics from IdentityUser class and will be the template for the final user will be filled in a future update.