Added Endpoints For The Rest Authentication And Authorization Library API
I added endpoints to the AuthLibraryAPI module that allow the user to utilize the authentication functionality that exist and the AuthLibrary module. As of now external identity providers have not been added. Also the authorization code is still missing and will be added in a later update(it is also not implemented in the AuthLibrary yet).