Added Integration Test For Authentication And Authentication Library
I added enough integration tests for the authlibrary.api module that test a typical flow of the authentication procedures. Some minor changes have been made in the authentication procedures for that reason and also in the api models. The method 'ChangeBasicAccountSettings' only does not have tests yet, because I am considering changing it or removing it. Further tests might be added in the future if that is deemed necessary.