Added External Identity Providers In The Auth Library And Massively Overhauled It
I added 3 methods, which will be used for the external authentication(OAuth 2 authentication). It must be said that without the AuthLibraryAPI this process is not possible, which is not ideal since the other parts could theoretically be used independently, but there is no other way of doing that. The main things that were changed were the returned codes that now the api returns, which make it more clear as to what error occured, so that the end user will have a better understanding of what happened. Another change that was made is that now the library has much more authority on what happens for a given request. That means that every endpoint of the authlibraryapi as it stands only calls one method of the authlibrary and it just interprets and returns the corresponding appropriate response. Additionally api key protection were added for the api, which drops unauthorized api calls and also I also added a basic rate limit mechanism that can help with DDOS attacks.