Added Endpoints For External Identity Providers
I added the necessary endpoints that will allow the user to sign up or sign in with the external identity provider of their choice. For now only Twitter/X and Google are configured, but I might add Facebook and Microsoft in the future. Because integration tests might be hard to add for the external login callback I decided to do some minor testing with postman and for now the endpoints(ExternalSignIn and ExternalSignInCallback) seems to work as intended, but further testing will be needed. Also changed some minor things like the DeleteAccount endpoint and its underlying library method and some minor things when it comes to logging in the auth library.