Added Integration Tests For The Added External Identity Provider Endpoints
I added tests for the 2 out of the 3 endpoints that I added in the previous commit. Those tests test a typical flow and some edge cases. The reason why I did not test the ExternalSignInCallback is because I have no idea how to do that. I could mock certain things, but it would defeat the purpose of integration tests. Hopefully the manual postman tests I did in the previous commit are enough to make sure that the endpoint does not have any logical errors.