Added Authorization Procedures In AuthLibrary
I added authorization procedures in the authlibrary in the form of the classes RoleManagementProcedures and AdminProcedures. RoleManagementProcedures takes care of managing roles and claims. The roles can be dynamic and an admin can create new roles that have claims, which are static. The authorization will happen throught claims and policies and thus this will give the ability to the admin to create custom roles with different permissions. The AdminProcedures allow an admin user to do CRUD operation on other users. Finally a HelperMethods class was also added, which contains common validation, authentication and authorization checks for the AuthLogic namespace. These changes are big and will need to be thoroughly tested in future commits, which means that a lot might change.