Added Authorization Endpoints In The Auth Library Api
I added the necessary authorization endpoints in the AuthLibraryApi module, that give the users the ability to access the utilities of the AdminProcedures and the RoleManagementProcedures that were added in the previous commit in the AuthLibrary. Some minor changes were also made in the name of the roles and in the configuration of some services.