Added Integrations Tests For The Authorization Endpoints
I added a ton of integrations tests for the authorization endpoints and massively overhauled the authorization system. The idea now is that there are elevated roles and non elevated roles, after authorization happens there are additional checks to make sure that a non elevated user can not edit or access an elevated role or user(depending on the protections they have). This should make sure that there will not be any security issues. As it stands a user can have multiple roles, but the system only cares for one of them, so I will need to maybe change that in the future or not allow a role to be added to a user if they already have 1. All the changes have been extensively tested and while there is room for refactoring there should not be many logical errors. The seeded claims have also slightly changed from the previous update for easier management, which admitedly limits customization.