Added The Inital Database Schema And DataAccess Operations For Category, Discount, Attribute, Variant And Product Entities In The DataLibrary
In this update a big chunk of the dataaccess operations, which are concerned with the eshop data of the application, were added. More specifically the CRUD operations were added for the category, attribute, discount, variant and product entities. As it stands certain things are not final. For example the way the images are handled(one-to-many with variants) might change. More importantly the way the updates work might change, but as it stands for now the current model allows for the update of only the immediate entity and its immediate connections with other entities. For example updateVariant would only update the columns of the variant entity itself and only its foreign keys, but nothing else. Other than that maybe some refactor needs to happen and improved logging with logging codes need to be added.