Added Endpoints For The Entities Categories, Products, Variants, Attributes, Images And Discounts
I added CRUD endpoints for the entities categories, products, variants, attributes, images and discounts. These endpoints might change in the future, but for now they support the basic operations, which were added in the previous commit with the appropriate error status codes if something goes wrong. The database schema changed with now the relationship between the variants and images is a many to many relationship. There were also some slight changes in the datatypes of the database and some unique constraints were added. The library also changed a bit with the major change being the addition of the ImageDataAccess class, which now implements its own CRUD operations on the image entity.