Added Integration Tests For The Entities Products, Variants, Discounts, Images, Categories And Attributes
I added a comprehensive amount of tests for the entities products, variants, discounts, images, categories and attributes. These tests take into account future entities that will be added to the database later, since those entities depend on specific functionality that some of the above entities have(needed to implement soft deleted mechanisms for example for variants, products and images). Additionaly a bridge entity ImageVariants was added explicitly for technical reasons. The schema of the database has also changed a little bit in order for the above to be implemented correctly(in general the fields for soft deleted procedures were added). The dataaccess also changed to accomodate these changes to. Another important thing to mention is that I have transitioned the library to use the partial update model, which means that the update endpoint allows the user to only send the information that they want to be updated and the other properties of the model will not be touched in this update.