Added Coupons Entity Modified Discounts Entity Added Endpoints To The API For Them And Tests For The Endpoints
I had already added a first implemetnation of the discounts entity in the previous commit, so in this commit I made sure that it now works better with the future orders entity, which will be added in a future commit. That means that it now has soft deleted/deactivated functionality. Additionally I added the Coupons entity and its bridge entity UserCoupons. I mistakenly also added in this update the endpoints for the the CRUD operations and the tests to them. Theoretically I should had split them, but I misread the issue and thus after this commit is done I will close the other 2 issues manually. So after this update the Discount Entity and Coupon Entity should be fully supported and they have also been thoroughly tested.