Added Orders, PaymentOptions, ShippingOptions Their Supporting Entities And Corresponding CRUD Operations For Them
I Added the Orders, PaymentOptions, ShippingOptions with their supporting entities(OrderItems, PaymentDetails, OrderAddresses)to the database schema. Most of the other database schema did not change much. A big change was made to the UserCoupons entity which now contains the soft deleted logic instead of the Coupons entity. Also the Products and Discounts entities no longer contain the ExistsInOrder column and they have to check through the variant property on whether or not they must be soft deleted or not. Other than the database schema changes, the CRUD operations for the 3 main entities(Orders, PaymentOptions and ShippingOptions) have been added throught the classes OrderDataAccess, PaymentOptionDataAccess and ShippingOptionDataAccess. As a final note it must be said that the OrderDataAccess could change in the future since it is the one CRUD logic class that contains the most complicated logic of all the CRUD logic classes(also invoice and receipt support might be added in the future).