Added API Endpoints For The CRUD Operations In Of The PaymentOption, ShippingOption and Order Entities
I added api endpoints for PaymentOption, ShippingOption and Order entities, which were added in the previous commit, to allow the user to access their corresponding CRUD operations in the dataaccess library. The necessary request models were added when they were needed to accomplish this goal. When it comes to the PaymentOption and ShippingOption the typical CRUD operations were added, but the Order entity also contains a change of state endpoint that is not typical. Finally some small changes were added to the schema to make these addition work as intented.