Added Integration Tests For The Order, PaymentOption And ShippingOption Endpoints
I added all the necessary tests for the Order, PaymentOption and ShippingOption endpoints, which were added in the previous commit. Most of the tests were added to test the Order endpoints, which contain the most complicated logic and potentionally more will be added for it in the future, if it is deemed necessary. The tests for the PaymentOption and ShippingOption should suffice. That being said with the addition of the Order entity all the entities that are connected with it(Variant, Image, PaymentOption, ShippingOption etc) could have some extra tests on their end, which test what happens when the order exists and they admin tries to remove them. I do not think this is necessary for now, but if errors start appearing I will also test for those scenarios.