Added Carts And CartItems Entities In The Database And Their Corresponding CRUD Operations In The DataLibrary
I added Carts and CartItems entities in the database and their corresponding CRUD operations in the CartDataAccess class in the DataLibrary module. Another small thing that changed is that now the userId column of the OrderEntity is not required considering the fact that a user might place an order as a guest. The endpoints for these CRUD operations will be added in the next update and thus things might change.