Added Endpoints In The API For The CRUD Operations That Correspond To The Cart And CartItems
I added all the necessary endpoints that allow the user to use the CRUD operations of the Cart and CartItems entities, which were added in the previous commit. To achieve this reason the CreateCartItemRequestModel.cs, CreateCartRequestModel.cs and the UpdateCartItemRequestModel.cs classes were also added and are used by the Post and Put endpoints. I also deleted a useless if statement that existed in all controllers in their delete endpoints. That was not a functional bug, but it helps with improved clarity. The test for the added endpoints will be added in the next update.