Added Integrations Tests For The Carts And CartItems Entities Endpoints
I added all the integration tests, that I considered necessary, for the endpoints of the Datalibrary API that correspond to the Carts and CartItems entities. Also the schema of the database had some very minor adjustments(userId got a unique constraint). Additionally the logic in the CartDataAccess now includes checks for insufficient variant stock, which might occur if the user picks an invalid quantity. The API has also been adjusted to accomodate those changes. This update should conclude the datalibrary commits, but I might return to add more logic and tests if that is deemed necessary.