Added Integration Tests For The Transaction API Endpoints
I added integration tests that test the endpoints CreateCheckOutSession and IssueRefund endpoints. These tests test most scenarios that can occur with the exception of the creation of a valid refund since I will need to do some difficult technical stuff in order to get the payment intent in my hands(maybe I will do that later). Another thing that I did not test are the webhooks, which will be really hard to test since I will have to trigger them somehow after the checkout or the payment intent is created, which seems to be really hard. That being said I have done some tests using postman to try them out and I will test them more in the tests of the gateway and in the end when I do my E2E testing. As it stands now I think the logic is mostly fine.