Added Authentication Endpoints In The Gateway Api
I added authentication endpoints in authentication controller of the gateway api that allow access to the front end client to the underlining authentication procedures of the Auth webservice. The Data webservice and the Email webservice also participate to some degree for those operations and thus the authentication controller utilizes them when he needs to(creation/deletion of user cart and coupons and confirmation/reset emails). For those purposes a new endpoint was added to the CouponController of the Data webservice, which allow for bulk removal of a user's coupons. The ForgotPassword endpoint in the Authentication controller of the auth webservice also now returns the user UserId since the Gateway api requires it for further processing. Retry mechanisms have already been added and rate limiting and api key authentication has also been added to the gateway API. One thing that might change in the future is the addition of a roleback mechanism, which will be useful if the execution of a gateway endpoint ends with an error after it has already made changes in a previous webservice. As of now there is no rollback mechanism due to the complexity it would add to the project.