Added Integration Tests For The Authentication Controller Of The Gateway API
I added all the necessary integration tests for the authentication controller of the gateway api. For that purpose certain changes were made in the the authentication controller when issues were found. Additionally I added Api key authentication and bandwidth limit to the email microservice. A small change was made to the authentication controller with it now invalidating previous access token(check on the email of the access token) upon email change. One important change is that health checks were added to all 4 microservices. The reason for this is state control. In some endpoints of the gateway api more than one microservice is called. If a system error appears in one of the microservices after some state change has occured in a previous microservice then I would have to make some sort of rollback. Since rollback logic is pretty complex and I have not implemented it yet the health checks check before any state changes occur that all the microservices that the endpoint depends upon are online and healthy(database, smtp server etc are ok). This health checks in combination with the retry logic that I have implemented should make sure that an invalid state should not happen. One small thing I need to add is that I could not make the external login work with the gateway api, which means that the client will have to bypass the gateway api for this one procedure. I am not happy with this and I might revisit it in the future, but as it stands I have not added the necessary logic for this endpoint.