Added Authorization Endpoints In The Gateway Api
I added all the necessary authorization endpoints that are needed to cover the admincontroller of the Auth microservice(CRUD operations that the admin can access). I also fixed a bug that made the endpoints of the gateway api to only sent health checks to the auth microservice. When I start working on the Data controllers of the gateway api I will also add some logic for the cart and the user coupons, but as it stands I will leave it as it is. On a final note maybe I should add in app notifications(for userupdate and other stuff). I am not certain about this yet.