Added Integration Tests For The Admin Controller In The Gateway API
I added all the necessary integration tests for the admin controller that was added in the previous commit to the gateway api. Those tests check the CRUD procedures that the admin can access through the gateway app(it communicates with the Auth Library Admin Controller for that purpose). When it comes to the code I just changed some return status codes that the admin controller of the auth microservice to notfound instead of bad request. The admincontroller of the gateway api also changed in some parts after issues were found through testing.