Added Role Management Endpoints In The Gateway Api
I added role management endpoints in the gateway api that correspond to each one of the endpoints of the RoleController of the AuthLibraryApi module. There is no much more to add since the tests to those endpoints will be added in the next update, but if I have planned correctly there should not be any further chagnes. One thing I noticed is that in case of getUsers as it stands the user roles are not retrieved, which might cause issues in user management. Yes for each user there is a way to get their role, but that will be a 1 + n requests and database queries, which is totally unacceptable. So if I decide that this needs to be done the auth microservice might change slightly, which might cause also the gateway api to also change, but as it stands I have not changed anything.