Added Integration Tests For The Role Management Endpoints In The Gateway Api
I added all the necessary integration tests for the role management endpoints that are needed to test all the previously added endpoints in the gatewayrolecontroller of the the gateway api. There were some minor adjustments to those endpoints after the testing, but nothing serious(now create role returns correctly 201 instead of 200 status code).