Added Category, Products, Variant, Image And Attribute Related Endpoints In The Gateway API
I added category, product, variant, image And attribute entitiy related endpoints in the gateway Api, which can be used to access the corresponding data endpoints of the data microservice for the those entities. When it comes to get operations those gateway endpoints call the corresponding data endpoint without any prior processing. That is not the case for create, update and delete, in which cases the user must be authorized and be in a role that contains the claim CanManageProducts(the gateway api endpoint checks for that). Other than that a folder restructuring occured for better management and 3 new seeded claims were added in the auth microservice(that might change in the future). An additional endpoint has been added in the authcontroller of the authlibraryapi, which is used by the create, update and delete endpoints that were added to check for correct user authentication and whether or not the user has the correct claim. All the above might change in the next commit after testing has been added to the new endpoints.