Added Integration Tests For The Category, Product, Variant, Image And Attribute Endpoints In The Gateway Api
I added all the necessary integration tests for the gateway, product, variant, image and attribute endpoint. Those tests in general mirror the ones that already exist in the integration tests for those entities in the authlibraryapi, but the also check for correct authentication and authorization since the gateway endpoints also add it to the mix. I also fixed an oversight(UpdateProductAsync in ProductDataAccess) in which if a variant through update was severed from the product is still remained in the database. Now it is correctly deleted and if it exists in an order it still remains connected to the product, but its deactivated flag is set to true. Other than that the gateway api endpoints that were added in the previous updated were changed a bit when bugs were found through testing.