Added Coupon And Discount Related Endpoints In The Gateway API
I added coupon and discount related endpoints in the gateway api, which can be used to access the corresponding discount and coupon endpoints of the data microservice. In general it was mostly calling those endpoints, but in most endpoints there is also added authentication and authorization. An important difference with previous authorization is that for the usercoupons instead of checking for CanManageProducts claim the endpoint checks for the CanManageUsers claim. The admin endpoints GetUserById, GetUserByEmail, and the authentication endpoint GetUserByAccessToken of the gateway api have also been updated and now they also retrieve the user coupons(maybe I should make this optional). Other than there is not much to add, but it should be noted that the endpoints will be tested in the next update, so certain things might change.