I Added Integration Tests For The Discount And Coupon Endpoints In The GatewayApi
I added all the necessary integration tests for the discount and coupon endpoints that exist in the gateway api(they were added in the previous update). Those tests are pretty similar to the tests that exist in the DiscountControllerTests and CouponControllerTests classes, but they also include authentication and authorization where it is needed. Also those tests use a real userId instead of a mock one and they also check that those coupons have been added properly to the user. More extensive tests might be added in the future for those checks, but for now I think those tests are suffecient.