Added PaymentOption, ShippingOption And Cart Related Endpoints In The Gateway API
I added paymentoption, shippingoption and cart related endpoints in the gateway api, which the client can use to access the underlying endpoints of the data microservice. Those endpoints in general just make a call to the corresponding endpoints in the data microservice, but a lot of them(create, update, delete) they also contain authentication and authorization. The cart endpoints also do contain authentication and they make sure that the cart or the cart item is being updated by the user that owns it otherwise they return forbiden(authorization). I also added and tested a new getVariantsBySkus endpoint(in both the data microservice and in the gateway api), which allow the client throught the gateway api to retrieve multiple variants by the skus. This I think is useful/efficient for asynchronous cart updates and I think I will use it there.