Added Integration Tests For PaymentOption, ShippingOption And Cart Endpoints In The Gateway API
I added all the necessary integration tests for the PaymentOption, ShippingOption and Cart endpoints in the gateway api. Those tests in general mirror those tests that exist for the endpoints in the data microservice, but they also test authentication and authorization scenarios. In the case of the cart tests the test that the user owns the cart and the cart item before any changes are made to it/them. Also some small changes were made to the previously added gateway endpoints after testing.