Added Order, CheckOut, Refund Related Endpoints In The Gateway Api
I added Order, CheckOut, Refund related endpoints in the gateway api. The updates that were made in this update probably have broken many tests and also a lot of refactoring took place. I did not have enought time to check and fix the things that are broken in this update and I will do this in the next update. Multiple new endpoints were added to support the added endpoints and also some bugs were fixed(updateOrder final price bug fixed). Again I did not have enough time in this update to do all that I needed to do, but the next update will wrap everything up.