Added Integration Tests For The Order, CheckOutSession Endpoints
I added endpoints that thoroughly test the Order endpoints and to a less degree the CheckOutSession endpoints. I did not add refund tests, because I ran out of time and because it would be really hard to do that. Maybe in a future update I could add more tests to check that things work 100%, but as it stands I will find any bugs left through the Front End tests.