Added Authentication And Authorization To FrontEnd
After this update all the necessary frontend and UI should be in place for the authentication of the user. The user now can signup, signin, logout, edit their account and delete their account. In future updates the user will be able to see their coupons, orders and alsoaddresses. External sign in logic has not been added yet and will be added in a later update. The logic has already been tested, but will fully be tested in the next update, in which the E2E will be added and will test that everything works as intended.