Added Product And Variant Management To FrontEnd
This update is huge and consists of multiple changes. Partial changes have been made to backend, but the most importants lie in the MVC module of the web app. In this I have added the all the necessary views, controllers, viewmodels, models and controller actions that facilitate all the CRUD operations for the Image, Variant and Product entities. The only thing that is left out due to time restrainsts is an ManageImage view in which the user can perform CRUD operations directly on the images themselves. Right now this is done using the CreateProduct and EditProduct views indirectly. Other than these changes has been massive refactorings to the javascript front end code. As it stands the filters have become more modular for reusability purposes and a reusable customImagePicker has been added. Most of the hardcoded code of the filters has been moved into the filtersRepeatedNonStaticCode.js file to avoid future repetition. All in all many general structures have been added which will be used in future commits. Filter checkbox logic has been also updated. Now a the or relationship of the different checkbox filter categories is correctly respected and the count of how many items is left is correctly calculated. Some minor bugs have also been fixed. Finally an important change is that now we Newtonsoft is the json parser of choice for our APIs. The reasoning behind this change is that there were cycle bugs that are now fixed, because of that change.