Added Discount And Coupon Management To The FrontEnd
This is a relatively smaller commit compared to the previous one. In this update all the view, controllers and models that are necessary for discount and coupon management have been added. Some minor bugs have also fixed and fixed responsive issues with the main header of manage views. Additionally I used the DataTable library to add some add the ability for a user to pick products and variant for the categories and the discounts accordingly more easily. This could have been done with a multi select, but I think that since there will be few categories and discounts and at the same there will be many products and variants it is not a good idea to do this with a multiselect. Multiselects are good when the options are few, but when they become hundreds then I dont think they are so ideal and my table implementation is probably more useful. This update has not been thoroughly tested and there might be bugs that I will have to fix in the next update.