Added First Part Of User Shopping Functionality To FrontEnd
I rewrote the ViewItems view, which now correctly shows the available items dynamically. I have also added the ViewItem view and the corresponding action to the Home Controller. The UI is mostly ready, but I need to implement the cart functionality, which probably will require me to use the browser's local storage and at the same time an AJAX request to the server for cart updates. Other than that this part is done, but I need 1 or 2 more days.